ad spot service, page , elements

+++ src/main/java/api/dto/rx/inventory/adspot/AdSpot.java
@@ -27,7 +27,7 @@ public class AdSpot {
     private Boolean floorPriceAutomated;
     private Boolean testMode;
     private Integer filterId;
-    private Integer floorPrice;
+    private Double floorPrice;
     private Integer ttl;
     private Integer positionId;
     private Integer prebidCacheEnabled;
